+++ backend/unithis/src/main/java/com/unithis/service/IChatRoomService.java
@@ -6,8 +6,8 @@ import com.unithis.model.ChatRoom;
 import com.unithis.model.ChatRoomRequest;
 
 public interface IChatRoomService {
-	public int createChatRoom(ChatRoom newRoom);
-	public List<ChatRoom> getJoinedRoomList(int user_id);
+	public long createChatRoom(ChatRoom newRoom);
+	public List<ChatRoom> getJoinedRoomList(long user_id);
 	public ChatRoom checkExistChatRoom(long user1_id, long user2_id);
 	public boolean updateReadTime(ChatRoomRequest chatroomReq);
 }
